src/main/python/pre_processors/asr/requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
beautifulsoup4==4.6.0
cachetools==2.0.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==1.9
dill==0.2.7
future==0.16.0
gapic-google-cloud-speech-v1==0.15.3
gevent==1.2.2
google==1.9.3
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.25.0
google-cloud-speech==0.26.0
google-gax==0.15.13
googleapis-common-protos==1.5.2
greenlet==0.4.12
grpcio==1.4.0
httplib2==0.10.3
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
jsocket==1.6
msgpack-python==0.4.8
netifaces==0.10.6
numpy==1.13.1
oauth2client==3.0.0
pika==0.10.0
ply==3.8
proto-google-cloud-speech-v1==0.15.3
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
PyAudio==0.2.11
pycparser==2.18
pyOpenSSL==17.1.0
pysolr==3.6.0
PyYAML==3.12
pyzmq==16.0.2
requests==2.18.1
rsa==3.4.2
service-identity==17.0.0
six==1.10.0
stt-watson==1.0.3
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
watson-developer-cloud==0.26.1
websocket-client==0.44.0
zope.interface==4.4.2
